Add styled heading
This improves the heading at the top of the commit list by applying some
colours and including information about the current status.

In order to do this the Rainbow gem was added, the first non stdlib
dependency. Therefore it was necessary to add a Gemfile as well.

Some of the existing messages have been tweaked and styled as well.

The Readline methodology for getting the window size turned out to be
unreliable to have switch to io/console instead which is still part of
the stdlib.