Add commits to show my / their changes when there are conflicts
These are a first draft of some commands to help provide information
about conflicts. It is expected that they will need some additional
tweaking as and when more conditions arrive. It would also be helpful to
simplify these by abstracting out helpers for things like normalising
paths relative to the git root.