Rewrite app to use new DSL declarative syntax
This ports the existing functionality over to the new DSL based
declarative syntax. All the commands have been copied over and redundant
classes have been removed meaning that the app functionality is now
fairly generic and reusable whilst the prepare-branch specific
functionality is now entirely configured through the DSL.

The possible exception to this are the autocomplete handlers which are
currently still defined in the Terminal class.