Prepare for release on Rubygems
The initial version will be released on Rubygems since this is the
simplest way to distribute a Ruby application.

This required moving the source code since the Rubygems convention is
that source lives in `/lib` and there is an file with the same
name as the gem that handles all the initial including.

This also meant that it was necessary to add a number of specific
requires in order to avoid constantly requiring the entire application
in tests.