Create OSX installation package
This uses travelling ruby to package the application into a
self-contained tarball that does not require a ruby installation on the
system to run.

It makes use of travelling ruby. It was necessary to perform some
modifications of their normal scripts in order to handle:

- installation to paths with a space in. Whilst this probably won't be
  required for the majority of users to run the application, my
  development version was at a path that contained a space in the name
  so it seemed better to try and fix the problem than hide it by moving
  the source to somewhere it didn't matter.

- dependencies are declared in the gemspec rather than the Gemfile.
  This is correct as the app can also be distributed via rubygems
  therefore it was necessary to also copy the gemspec and version file
  along with the Gemfiles when generating the bundle and package.

Whilst the Rakefile is set up for linux and OSX distributions only the
OSX distribution has been created. This is because the first target will
be Homebrew. The rest of the distributions will be created when we are
ready to add them to the appropriate package manager(s).